requirements.txt
simplejson
django==1.6
django-tastypie